[datakit] Move intruder modules under cluster/
intruder.py and test_intruder.py now live in experiments/datakit/cluster/,
alongside the other clustering experiments. Update the import and the caplog
logger name to the new module path.
